+++ .flake8
@@ -1,0 +1,4 @@
+[flake8]
+max-line-length=88
+ignore=
+    E741 # ambiguous variable name
