+++ .isort.cfg
@@ -8,3 +8,6 @@ known_first_party=ayeauth
 sections=FUTURE,STDLIB,COMPAT,TEST,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 default_section=THIRDPARTY
 multi_line_output=3
+include_trailing_comma=True
+force_grid_wrap=0
+use_parentheses=True
