+++ .pre-commit-config.yaml
@@ -17,15 +17,14 @@ repos:
   - id: mixed-line-ending
     args: ['--fix=no']
   - id: flake8
-    args: ['--max-line-length=88']  # default of Black
+
+- repo: https://github.com/pre-commit/mirrors-isort
+  rev: v4.3.4
+  hooks:
+  - id: isort
 
 - repo: https://github.com/psf/black
   rev: stable
   hooks:
     - id: black
       language_version: python3.8
-
-- repo: https://github.com/pre-commit/mirrors-isort
-  rev: v4.3.4
-  hooks:
-  - id: isort
