+++ src/ayeauth/__init__.py
@@ -1,5 +1,3 @@
-import uuid
-
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager
@@ -11,10 +9,6 @@ app = None
 db = SQLAlchemy()
 lm = LoginManager()
 pr = Principal()
-
-
-def _get_uuid():
-    return str(uuid.uuid4())
 
 
 def create():
