+++ src/ayeauth/application/forms.py
@@ -1,24 +1,29 @@
 from flask_wtf import FlaskForm
-from wtforms import BooleanField, StringField, SubmitField, validators, SelectField, FieldList, FormField, TextAreaField
+from wtforms import (
+    FieldList,
+    FormField,
+    SelectField,
+    StringField,
+    SubmitField,
+    TextAreaField,
+    validators,
+)
 
-from ayeauth import _get_uuid
-from ayeauth.models.application import Application
-from ayeauth.models.scope import ScopeAccess, Scope
+from ayeauth.models import _get_uuid
+from ayeauth.models.scope import Scope, ScopeAccess
 
 
-def generate_scope_select_fields():
-    scopes = Scope.query.all()
-    scope_field_list = []
+def scope_select_fields():
+    l = []
+    for scope in Scope.query.all():
+        ssf = ScopeSelectField()
+        ssf.id = _get_uuid()
+        ssf.scope.label = scope.name
+        ssf.scope.description = scope.description
+        ssf.scope.choices = [scope_access.value for scope_access in ScopeAccess]
+        l.append(ssf)
 
-    for scope in scopes:
-        scope_field = ScopeSelectField()
-        scope_field.id = _get_uuid()
-        scope_field.scope.label = scope.name
-        scope_field.scope.description = scope.description
-        scope_field.scope.choices = [scope_access.value for scope_access in ScopeAccess]
-        scope_field_list.append(scope_field)
-
-    return scope_field_list
+    return l
 
 
 class ScopeSelectField(FlaskForm):
@@ -29,7 +34,7 @@ class ApplicationForm(FlaskForm):
     name = StringField("Name", [validators.InputRequired()])
     description = TextAreaField("Description")
     callback_url = StringField("Callback URL", [validators.InputRequired()])
-    scope_list = FieldList(FormField(ScopeSelectField))
+    scope_select_fields = FieldList(FormField(ScopeSelectField))
     submit = SubmitField("Register Application")
 
     def validate(self):
