+++ src/ayeauth/application/routes.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, render_template
 from flask_login import current_user, login_required
 
-from ayeauth.application.forms import ApplicationForm, generate_scope_select_fields
+from ayeauth.application.forms import ApplicationForm, scope_select_fields
 
 application_bp = Blueprint(
     "application_bp",
@@ -16,25 +16,9 @@ application_bp = Blueprint(
 @login_required
 def create():
     form = ApplicationForm()
-    form.scope_list = generate_scope_select_fields()
+    form.scope_select_fields = scope_select_fields()
 
     if form.validate_on_submit():
         pass
 
     return render_template("create.html", form=form, user=current_user)
-
-
-@application_bp.route("/test")
-def test():
-    from ayeauth import db
-    from ayeauth.models.scope import Scope
-
-    a = Scope("email", "")
-    b = Scope("username", "")
-    c = Scope("dateofbirth", "")
-    db.session.add(a)
-    db.session.add(b)
-    db.session.add(c)
-    db.session.commit()
-
-    return {}
