+++ src/ayeauth/application/templates/create.html
@@ -29,16 +29,16 @@
             <h2>Permissions</h2>
             <br>
 
-            {% for scope in form.scope_list %}
+            {% for ssf in form.scope_select_fields %}
             <div class="form-group">
                 <div class="row">
                     <div class="col">
-                        <p><strong>{{ scope.scope.label }}</strong><br>
-                        <small>{{ scope.scope.description }}</small></p>
+                        <p><strong>{{ ssf.scope.label }}</strong><br>
+                        <small>{{ ssf.scope.description }}</small></p>
                     </div>
 
                     <div class="col float-right">
-                        {{ scope.scope(class_="form-control") }}
+                        {{ ssf.scope(class_="form-control") }}
                     </div>
                 </div>
             </div>
