+++ src/ayeauth/models/application.py
@@ -1,5 +1,5 @@
-from ayeauth import db, _get_uuid
-from ayeauth.models import BaseModel
+from ayeauth import db
+from ayeauth.models import BaseModel, _get_uuid
 
 
 class Application(BaseModel):
@@ -9,7 +9,11 @@ class Application(BaseModel):
     description = db.Column(db.String())
     callback_url = db.Column(db.String(), nullable=False)
     client_id = db.Column(db.String(36), nullable=False, default=_get_uuid)
-    scopes = db.relationship("Scope", secondary="application_scopes", backref=db.backref("applications", lazy="dynamic"))
+    scopes = db.relationship(
+        "Scope",
+        secondary="application_scopes",
+        backref=db.backref("applications", lazy="dynamic"),
+    )
 
     def __init__(self, name, description, callback_url):
         super(Application, self).__init__()
